✏️ rename services file, update routes paths
