+++ frontend/src/routes/routesConfig.tsx
@@ -3,9 +3,9 @@ import { Register } from "../auth/register";
 import { Newpassword } from "../components/NewPassword/newpassword";
 import { Recoverpassword } from "../components/RecoverPassword/recoverpassword";
 import { Contact } from "../shared/contact";
-import { Gallery } from "../shared/gallery";
+import { Gallery } from "../shared/gallery/gallery";
 import { Homepage } from "../shared/homepage";
-import { Servicios } from "../shared/services/servicios";
+import { Servicios } from "../shared/services/services";
 import { Cart } from "../shared/shop/shoppingCart";
 
 export const routes = [
